nixos/networking: don't add hostname / FQDN entires to `/etc/hosts`
PR #362132 [1] removed multiple `/etc/hosts` entires mapping to the
same IPv6 loopback address (`::1`). However, a mapping from a
machine's hostname and FQDN still exists for a second IPv4 loopback
address (`127.0.0.2`).

This causes issues when using systemd's resolved: when any
`/etc/hosts` entry exists for a domain name being queried,
systemd-resolved only returns entires from the `/etc/hosts` database,
and no other records that can be obtained from DNS. In practice, this
means that a host trying to resolve its own FQDN (which is further
configured in the system's NixOS configuration through the
`networking.hostName` and `networking.domain` settings) will only
return its IPv4 loopback IP (`127.0.0.2`), but _no_ IPv6 address,
regardless of whether a AAAA-record exists.

    [root@myserver:~]# cat /etc/hosts
    127.0.0.1 localhost
    ::1 localhost
    127.0.0.2 myserver.example.org

    [root@myserver:~]# resolvectl query myserver.example.org
    myserver.example.org: 127.0.0.2

    -- Information acquired via protocol DNS in 6.1ms.
    -- Data is authenticated: yes; Data was acquired via local or encrypted transport: yes
    -- Data from: synthetic

This pull request removes such default hostname or FQDN hosts-entries
entirely. This is similar to the default configurations on AlmaLinux
or Fedora systems:

    root@myserver:~# cat /etc/hosts
    # Loopback entries; do not change.
    # For historical reasons, localhost precedes localhost.localdomain:
    127.0.0.1   localhost localhost.localdomain localhost4 localhost4.localdomain4
    ::1         localhost localhost.localdomain localhost6 localhost6.localdomain6
    # See hosts(5) for proper format and other examples:
    # 192.168.1.10 foo.example.org foo
    # 192.168.1.13 bar.example.org bar

It does not prevent users from defining any custom such entires for
the machine's public IPs, similar to the comment in the file above. We
don't have a good idea to determine what the proper public IPv4 and
IPv6 addresses are for a machine's FQDN, and as such this decision and
responsibility should be on the user.

[1]: https://github.com/NixOS/nixpkgs/pull/362132